Moved internal functions into their own file.
Added missing test.
