+++ includes/RhythmGameUtilities/RhythmGameUtilities.cpp
@@ -2,4 +2,6 @@
 
 #include "Audio.hpp"
 #include "Parsers.hpp"
+#include "ParsersInternal.hpp"
 #include "Utilities.hpp"
+#include "UtilitiesInternal.hpp"
